add the comment input
