+++ src/Components/Notes/NoteCardCreate.jsx
@@ -10,7 +10,7 @@ import InputBase from '@mui/material/InputBase'
 import Stack from '@mui/material/Stack'
 import {TooltipIconButton} from '../Buttons'
 import useStore from '../../store/useStore'
-import {createIssue} from '../../utils/GitHub'
+import {createIssue, getIssueComments, createComment} from '../../utils/GitHub'
 import {assertStringNotEmpty} from '../../utils/assert'
 import CheckBoxOutlinedIcon from '@mui/icons-material/CheckBoxOutlined'
 
@@ -26,10 +26,12 @@ export default function NoteCardCreate({
   username = '',
   avatarUrl = '',
   isNote = true,
+  noteNumber = '',
 }) {
   const {user, isAuthenticated} = useAuth0()
   const accessToken = useStore((state) => state.accessToken)
   const repository = useStore((state) => state.repository)
+  const setComments = useStore((state) => state.setComments)
   const toggleIsCreateNoteActive = useStore((state) => state.toggleIsCreateNoteActive)
   const [title, setTitle] = useState('')
   const [body, setBody] = useState(null)
@@ -51,7 +53,45 @@ export default function NoteCardCreate({
     toggleIsCreateNoteActive()
   }
 
-  const submitEnabled = title !== null && title !== ''
+  const fetchComments = async () => {
+    const newComments = []
+    const commentArr = await getIssueComments(repository, noteNumber, accessToken)
+
+    if (commentArr) {
+      commentArr.map((comment) => {
+        newComments.push({
+          id: comment.id,
+          body: comment.body,
+          date: comment.created_at,
+          username: comment.user.login,
+          avatarUrl: comment.user.avatar_url,
+          synched: true,
+        })
+      })
+    }
+    setComments(newComments)
+  }
+
+
+  /**
+   * create new comment
+   *
+   * @param {string} repository
+   * @param {string} accessToken
+   * @param {number} commentId
+   * @return {object} return github return object
+   */
+  async function createNewComment() {
+    assertStringNotEmpty(body)
+    const commentPayload = {
+      body: body || '',
+    }
+    await createComment(repository, noteNumber, commentPayload, accessToken)
+    setBody('')
+    fetchComments()
+  }
+
+  const submitEnabled = (title !== null && title !== '') || (!isNote && body !== null && body !== '')
   return (
     <Card
       elevation={1}
@@ -89,7 +129,7 @@ export default function NoteCardCreate({
             onChange={(event) => setBody(event.target.value)}
             fullWidth
             multiline
-            placeholder={'Note Body'}
+            placeholder={isNote ? 'Note Body' : 'Comment Body' }
             inputProps={{maxLength: 256}}
           />
         </Box>
@@ -103,9 +143,7 @@ export default function NoteCardCreate({
         >
           <TooltipIconButton
             title='Submit'
-            onClick={async () => {
-              await createNote()
-            }}
+            onClick={isNote ? createNote : createNewComment}
             icon={<CheckBoxOutlinedIcon/>}
             enabled={submitEnabled}
             size='small'
